# Increase HttpClientHandler.MaxConnectionsPerServer to 64 to improve PM UI performance in Visual Studio