--- a/src/NuGet.Core/NuGet.Protocol/HttpSource/HttpSourceResourceProvider.cs
+++ b/src/NuGet.Core/NuGet.Protocol/HttpSource/HttpSourceResourceProvider.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Concurrent;
 using System.Diagnostics;
+using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
 using NuGet.Common;
@@ -49,7 +50,8 @@ namespace NuGet.Protocol
                     throttle = SemaphoreSlimThrottle.CreateSemaphoreThrottle(source.PackageSource.MaxHttpRequestsPerSource);
                 }
 #if IS_DESKTOP
-                else if (RuntimeEnvironmentHelper.IsWindows)
+                else if (RuntimeEnvironmentHelper.IsWindows
+                    && ServicePointManager.DefaultConnectionLimit == ServicePointManager.DefaultPersistentConnectionLimit)
                 {
                     source.PackageSource.MaxHttpRequestsPerSource = 64;
                 }