--- a/test/NuGet.Core.Tests/NuGet.Protocol.Tests/HttpSourceResourceProviderTests.cs
+++ b/test/NuGet.Core.Tests/NuGet.Protocol.Tests/HttpSourceResourceProviderTests.cs
@@ -1,6 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System.Threading.Tasks;
 using NuGet.Configuration;
 using NuGet.Protocol.Core.Types;
 using NuGet.Test.Utility;
@@ -10,39 +11,53 @@ namespace NuGet.Protocol.Tests
 {
     public class HttpSourceResourceProviderTests
     {
-
 #if IS_DESKTOP
         [PlatformFact(Platform.Windows)]
-        public void MaxHttpRequestsPerSourceIsToDefaultValue_Success()
+        public async Task WhenMaxHttpRequestsPerSourceIsNotConfiguredThenItsValueIsSetToDefault_SuccessAsync()
         {
             // Arrange
             var packageSource = new PackageSource("https://contoso.com/v3/index.json");
             var sourceRepository = new SourceRepository(packageSource, new[] { new HttpSourceResourceProvider() });
 
             // Act
-            HttpSourceResource httpSourceResource = sourceRepository.GetResource<HttpSourceResource>();
+            HttpSourceResource httpSourceResource = await sourceRepository.GetResourceAsync<HttpSourceResource>();
 
             // Assert
             Assert.NotNull(httpSourceResource);
             Assert.Equal(64, sourceRepository.PackageSource.MaxHttpRequestsPerSource);
         }
-#endif
-
-#if IS_CORECLR
+#elif IS_CORECLR
         [Fact]
-        public void MaxHttpRequestsPerSourceIsSetToDefaultValue_Success()
+        public async Task WhenMaxHttpRequestsPerSourceIsNotConfiguredThenItsValueWillNotBeUpdated_SuccessAsync()
         {
             // Arrange
             var packageSource = new PackageSource("https://contoso.com/v3/index.json");
             var sourceRepository = new SourceRepository(packageSource, new[] { new HttpSourceResourceProvider() });
 
             // Act
-            HttpSourceResource httpSourceResource = sourceRepository.GetResource<HttpSourceResource>();
+            HttpSourceResource httpSourceResource = await sourceRepository.GetResourceAsync<HttpSourceResource>();
 
             // Assert
             Assert.NotNull(httpSourceResource);
             Assert.Equal(0, sourceRepository.PackageSource.MaxHttpRequestsPerSource);
         }
 #endif
+
+        [PlatformTheory]
+        [InlineData(128)]
+        [InlineData(256)]
+        public async Task WhenMaxHttpRequestsPerSourceIsConfiguredThenItsValueWillNotBeUpdated_SuccessAsync(int maxHttpRequestsPerSource)
+        {
+            // Arrange
+            var packageSource = new PackageSource("https://contoso.com/v3/index.json") { MaxHttpRequestsPerSource = maxHttpRequestsPerSource };
+            var sourceRepository = new SourceRepository(packageSource, new[] { new HttpSourceResourceProvider() });
+
+            // Act
+            HttpSourceResource httpSourceResource = await sourceRepository.GetResourceAsync<HttpSourceResource>();
+
+            // Assert
+            Assert.NotNull(httpSourceResource);
+            Assert.Equal(maxHttpRequestsPerSource, sourceRepository.PackageSource.MaxHttpRequestsPerSource);
+        }
     }
 }