<!-- DO NOT MODIFY OR DELETE THIS TEMPLATE. IT IS USED IN AUTOMATION. -->
## Bug

<!-- Search https://github.com/NuGet/Home/issues, and create one if you can't find a suitable issue. -->
<!-- Paste the full link, like https://github.com/nuget/home/issues/1000. GitHub will render is neatly. -->
Fixes: https://github.com/NuGet/Home/issues/11923

Regression? Last working version:

## Description
<!-- Add details about the fix. Include any information that would help the maintainer review this change effective. -->

Thank you @mjolka for identifying this performance improvement and creating the linked issue. For context, [`HttpClientHandler.MaxConnectionsPerServer`](https://docs.microsoft.com/dotnet/api/system.net.http.httpclienthandler.maxconnectionsperserver?view=net-6.0) is by default `2` in .NET framework but the default value in .NET Core is `Int32.MaxValue`. Reference https://docs.microsoft.com/en-gb/archive/blogs/timomta/controlling-the-number-of-outgoing-connections-from-httpclient-net-core-or-full-framework. The performance improvements of Visual Studio PM UI mentioned in the linked issue description were great. Hence, in this PR I propose to increase `HttpClientHandler.MaxConnectionsPerServer` to `64` in `.NET Framework` code paths but don't change this property in `.NET` code paths because of the default value i.e., `Int32.MaxValue`. NuGet.exe on Windows [sets max connections limit per endpoint to 64](https://github.com/NuGet/NuGet.Client/blob/a56c42dbdd901c0c611eece400ce8565fffdfdf7/src/NuGet.Core/NuGet.Common/NetworkProtocolUtility.cs#L19). Hence, I used the same value `64` in this PR. The proposed changes also allow customers to tweak `HttpClientHandler.MaxConnectionsPerServer` by adding [`maxHttpRequestsPerSource`](https://github.com/NuGet/Home/issues/4538) in [NuGet.Config](https://docs.microsoft.com/en-us/nuget/reference/nuget-config-file) file.

I asked `.NET Networking team` if there would be any issues upgrading `MaxConnectionsPerServer` value in .NET Framework code paths. The response received so far from the team seems to be positive https://github.com/orgs/dotnet/teams/ncl/discussions/1. Hence, I went ahead with the changes. Thanks to @drewgillies for taking the time to test a private build with this fix. He mentioned that this fix made a lot of difference to the usability and performance of PM UI in Visual Studio.

Step | With preinstalled PMUI | With PMUI installed from a private build with this fix
-- | -- | --
Navigate to PMUI for solution, landing on Installed tab | 35s | 10s
Change to Browse tab | 36s | 13s
Change back to Installed tab | 15s | 3s
Change to Updates tab | 27s | 6s

~I don't have numbers handy right now, but I noticed great performance improvements in Visual Studio solution restore scenario. I will run a quick test and post the results here.~

EDIT: I captured few metrics. This PR reduced [OrchardCore](https://github.com/OrchardCMS/OrchardCore/commit/c4e37bdd6f63f3a5acdb018e101436cc5fd90a1f) VS Solution restore by **10 seconds**.

- VS Solution restore time without this fix.

MaxConnectionsPerSever | MaxHttpRequestsPerSource | Time for Solution restore
-- | -- | --
2 (Default) | Not set | **45 seconds**
2  (Default)| 64 | 42 seconds
2  (Default)| 128 | 39 seconds

- VS Solution restore time with this fix.

MaxConnectionsPerSever | MaxHttpRequestsPerSource | Time for Solution restore
-- | -- | --
64 | Default (64) | **35 seconds**
128 | 128 | 34 seconds

- `dotnet restore (7.0.100-preview.6)` for OrchardCore solution in my local machine took `32 seconds`. This number is close to VS solution restore scenario for default `MaxConnectionsPerSever` proposed in this PR which is `64`.

## PR Checklist

- [x] PR has a meaningful title
- [x] PR has a linked issue.
- [x] Described changes

- **Tests**
  - [x] Automated tests added

- **Documentation**
  <!-- Please link the PR/issue if appropriate -->
  - [x] N/A