@nkolev92 2022-09-15
> I think this PR also fixes https://github.com/NuGet/Home/issues/11638 issue. I can't reproduce this issue by installing VSIX from this PR build.

Personally, I have no problem with you marking that issue as being fixed as well.

If anyone from @NuGet/nuget-client disagrees, they should chime in :)