@zivkan 2022-09-15
> > I think this PR also fixes https://github.com/NuGet/Home/issues/11638 issue. I can't reproduce this issue by installing VSIX from this PR build.

> If anyone from https://github.com/orgs/NuGet/teams/nuget-client disagrees, they should chime in

I disagree. If VS was hanging when making network calls, that needs to be fixed by not wating for the network to finish on the UI thread. So, it might have been fixed by https://github.com/NuGet/NuGet.Client/pull/4739.  If it wasn't, then the duration of the UI delay might be reduced by this PR, but it doesn't fix network calls being made on the UI thread.