@kartheekp-ms 2022-09-16
> > > > I think this PR also fixes [NuGet/Home#11638](https://github.com/NuGet/Home/issues/11638) issue. I can't reproduce this issue by installing VSIX from this PR build.
> > 
> > 
> > > If anyone from https://github.com/orgs/NuGet/teams/nuget-client disagrees, they should chime in
> > 
> > 
> > I disagree. If VS was hanging when making network calls, that needs to be fixed by not wating for the network to finish on the UI thread. So, it might have been fixed by #4739. If it wasn't, then the duration of the UI delay might be reduced by this PR, but it doesn't fix network calls being made on the UI thread.
> 
> I actually agree with you :D I didn't read that issue in too much detail. It was actually a real hang, with UI freezes and all. This change simply can't fix that.

@zivkan @nkolev92 I agree with both of you :). I am unable to reproduce this issue https://github.com/NuGet/Home/issues/11638 even on the internal dogfooding version. Given that @jebriede fixed this UI delay in https://github.com/NuGet/NuGet.Client/pull/4739, I would let him close https://github.com/NuGet/Home/issues/11638 this issue as well.