@nkolev92 2022-09-16
I think the problem in https://github.com/NuGet/Home/issues/11638 is that we don't understand what exact codepath no?
Like it could be the same root cause as the one fixed in #4739, but we probably can't know now.