@kartheekp-ms 2022-09-21
I noticed typos in the test names I added in this PR. Hence pushed a new commit updating the test names and added a comment to explain why discard is used to ignore a return value in one of the test cases.