@kartheekp-ms 2022-09-26
I have manually queued a PR build and updated the variable to run end to end/apex tests on VS 17.3. CI checks passed for this PR. I am planning to merge this PR today.